# openapi.yml
openapi: '3.0.3'
info:
  title: MpesaFlow API
  version: '1.0'
  description: API for managing MpesaFlow transactions, applications, and API keys
servers:
  - url: https://api.mpesaflow.com
    description: Production server
  - url: https://sandbox-api.mpesaflow.com
    description: Sandbox server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
paths:
...
  /apps/{appId}/api-keys/create:
    post:
      summary: Create a new API key for an application
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: appId
          schema:
            type: string
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                keyName:
                  type: string
      responses:
        '201':
          description: API key created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  apiKeyId:
                    type: string
                  message:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /apps/{appId}/api-keys/{apiKeyId}:
    delete:
      summary: Delete an API key
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: appId
          schema:
            type: string
          required: true
        - in: path
          name: apiKeyId
          schema:
            type: string
          required: true
      responses:
        '200':
          description: API key deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
security:
  - BearerAuth: []